requirements.txt
streamlit==0.49.0
vega-datasets==0.7.0
altair==3.2.0
Pillow==6.2.1
scikit-learn==0.21.3
seaborn==0.9.0
appdirs==1.4.1
click==7.0
Flask==0.12
jupyter==1.0.0
lxml==4.4.1
matplotlib==3.1.1
pandas==0.25.2
gunicorn==19.6.0
itsdangerous==0.24
Jinja2==2.9.5
MarkupSafe==0.23
packaging==16.8
protobuf==3.6.1
pyparsing==2.1.10
six==1.11.0
tensorflow==2.0.0a0
Werkzeug==0.11.15